move findFileByName to utils
