+++ src/main.ts
@@ -2,55 +2,7 @@ import * as core from '@actions/core'
 import * as github from '@actions/github'
 import * as fs from 'fs'
 import * as path from 'path'
-
-/**
- * Recursively searches for a file with a specific name in a given directory.
- * @param directory The directory to start the search from.
- * @param fileName The name of the file to search for.
- * @returns The path of the found file or null if the file is not found.
- */
-function findFileByName(directory: string, fileName: string): string | null {
-  /**
-   * Recursive helper function to search for the file.
-   * @param currentDir The current directory to search.
-   */
-  function search(currentDir: string): string | null {
-    const items = fs.readdirSync(currentDir)
-
-    // Go over everything in the current directory
-    for (const item of items) {
-      const itemPath = path.join(currentDir, item)
-
-      // Check if the current item (file or dir) is what we are looking for
-      if (item === path.basename(fileName)) {
-        // File found in the current directory
-        return itemPath
-      }
-
-      // If the item is a directory descend into it
-      const isDirectory = fs.statSync(itemPath).isDirectory()
-      if (isDirectory) {
-        // Recursive call for subdirectories
-        const result = search(itemPath)
-        if (result !== null) {
-          // File found, propagate the result up
-          return result
-        }
-      }
-    }
-    // File not found in this directory or its subdirectories
-    return null
-  }
-
-  // Actual function body
-  try {
-    return search(directory)
-  } catch (error) {
-    if (error instanceof Error)
-      console.error(`Error while searching for a file: ${error.message}`)
-    return null
-  }
-}
+import { findFileByName } from './utils'
 
 /**
  * Checks that if any tagged source file was changed, its corresponding doc file was changed too.
@@ -68,7 +20,7 @@ function checkDocumentation(userdocs: string[], changes: string[]): number {
     // Get list of file tags. (All words that end with .txt, .h, or .cpp without full paths)
     const fileContent = fs.readFileSync(docfile, 'utf-8')
     let fileTags = Array.from(
-      fileContent.match(/[^\/\s]+\.(h|cpp|txt)\b/g) || []
+      fileContent.match(/[^/\s]+\.(h|cpp|txt)\b/g) || []
     )
     // Get list of directory tags (All strings that end on '/' after "Related ...Folders"; this split is case-insensitive:w
     const directoryTags: string[] = Array.from(
@@ -78,7 +30,7 @@ function checkDocumentation(userdocs: string[], changes: string[]): number {
     )
 
     const docfileHasChanges = changesBasenames.includes(path.basename(docfile))
-    let unknownTags: string[] = []
+    const unknownTags: string[] = []
     // append the content of all directories to the tags
     for (const tag of [...directoryTags]) {
       // If the path in the tag doesn't exist, it's an error
@@ -154,7 +106,7 @@ export async function run(): Promise<void> {
     const changedFiles = response.data.map(file => file.filename)
     core.info(`changed files: ${changedFiles}`)
 
-    const exitCode = checkDocumentation(docFiles, changedFiles)
+    /*const exitCode =*/ checkDocumentation(docFiles, changedFiles)
 
     // Set outputs for other workflow steps to use
     // TODO: use exitCode
